requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
django-configurations==2.1
django-debug-toolbar==1.11
django-filter==2.1.0
django-guardian==2.0.0
django-model-utils==3.1.2
django-unixtimestampfield==0.3.9
djangorestframework==3.9.4
entrypoints==0.3
flake8==3.7.7
gunicorn==19.9.0
idna==2.8
mccabe==0.6.1
mysqlclient==1.4.2.post1
pep8==1.7.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
pymongo==3.8.0
# django-allauth==0.39.1
djangorestframework-jwt==1.11.0
pyjwt==1.7.1
django-countries==5.3.3